Add screens for user authentication and settings management
- Implement ForgetPasswordScreen for password recovery.
- Create HomeScreen for displaying VPN status and server selection.
- Add VPNHomeScreen for when the user is not connected.
- Develop OnboardingScreen to guide new users.
- Introduce OTPVerificationScreen for verifying user identity.
- Implement ResetPasswordScreen for changing user passwords.
- Create ServiceScreen to showcase app services.
- Add SplashScreen for initial app loading animation.
- Implement SubscriptionScreen for managing user subscriptions.
- Create TermsAndConditionsScreen to display legal information.
- Add SettingsController and SettingsView for user preferences management.
