+++ lib/data/models/vpn_configuration.dart
