+++ lib/data/models/vpn_info.dart
