+++ lib/modules/screens/reset_password_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:vpn_app/screens/subscription_screen.dart';
-import 'package:vpn_app/widgets/custom_app_bar.dart';
-import 'package:vpn_app/widgets/custom_elevated_button.dart';
-import 'package:vpn_app/widgets/custom_text_field.dart';
+import 'package:vpn_app/modules/screens/subscription_screen.dart';
+import 'package:vpn_app/core/widgets/custom_app_bar.dart';
+import 'package:vpn_app/core/widgets/custom_elevated_button.dart';
+import 'package:vpn_app/core/widgets/custom_text_field.dart';
 
 class ResetPasswordScreen extends StatelessWidget {
   const ResetPasswordScreen({super.key});
