--- lib/screens/settings_screen.dart
@@ -1,145 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:vpn_app/widgets/custom_app_bar.dart';
-
-class SettingsScreen extends StatefulWidget {
-  const SettingsScreen({super.key});
-
-  @override
-  State<SettingsScreen> createState() => _SettingsScreenState();
-}
-
-class _SettingsScreenState extends State<SettingsScreen> {
-  bool isDataMode = true;
-  bool isNotifications = false;
-  bool isDarkMode = true;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.black,
-      appBar: CustomAppBar(
-        title: 'Settings',
-        onBackPressed: () {
-          Get.back();
-        },
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(20),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            _buildSwitchTile(
-                'Data Mode', 'VPN is also works on data mode', isDataMode,
-                (value) {
-              setState(() => isDataMode = value);
-            }),
-            _buildSwitchTile(
-                'Notifications',
-                'We will send you a notifications on your phone',
-                isNotifications, (value) {
-              setState(() => isNotifications = value);
-            }),
-            _buildSwitchTile(
-                'Dark Mode',
-                'Enable dark mode to make your eye comfortable',
-                isDarkMode, (value) {
-              setState(() => isDarkMode = value);
-            }),
-            const SizedBox(height: 20),
-            _buildSectionTitle('App Informations'),
-            _buildInfoTile('App Security',
-                'Enable dark mode to make your eye comfortable'),
-            _buildInfoTile('Privacy & Policy',
-                "A privacy policy is a statement or legal document that discloses some or all of the ways a party gathers, uses, discloses, and manages a customer or client's data."),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildSwitchTile(
-      String title, String subtitle, bool value, Function(bool) onChanged) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 10),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style: const TextStyle(
-                  color: Colors.white,
-                  fontSize: 16,
-                  fontFamily: 'Gilroy',
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              Text(
-                subtitle,
-                style: TextStyle(
-                  color: Colors.grey[500],
-                  fontSize: 12,
-                  fontFamily: 'Gilroy',
-                ),
-              ),
-            ],
-          ),
-          Switch(
-            value: value,
-            onChanged: onChanged,
-            activeColor: Colors.white,
-            activeTrackColor: const Color(0xff34c759),
-            inactiveTrackColor: const Color(0xff131314),
-            inactiveThumbColor: Colors.white,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSectionTitle(String title) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 10),
-      child: Text(
-        title,
-        style: const TextStyle(
-          color: Colors.white,
-          fontSize: 20,
-          fontFamily: 'Gilroy',
-          fontWeight: FontWeight.bold,
-        ),
-      ),
-    );
-  }
-
-  Widget _buildInfoTile(String title, String subtitle) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 15),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            title,
-            style: const TextStyle(
-              color: Colors.white,
-              fontSize: 16,
-              fontFamily: 'Gilroy',
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-          Text(
-            subtitle,
-            style: TextStyle(
-              color: Colors.grey[500],
-              fontSize: 12,
-              fontFamily: 'Gilroy',
-             ),
-          ),
-        ],
-      ),
-    );
-  }
-}
