+++ pubspec.yaml
@@ -35,7 +35,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
-  get: ^4.6.6
+  get: ^4.7.2
   hive: ^2.2.3
   hive_flutter: ^1.1.0
   http: ^1.2.2
